Fixed(?) Loading of settings
I'm considering removing the settings, because that part of the plugin keeps crashing and I can't tell why. Apparently the value returned by the settings object is not what it should be, an int. Apparently, sometimes it can't be converted to an int, so then I have to fall back to a default. I *think* it should work now, but I'm not sure because it always takes some time to fail and I haven't pinned down yet what causes the problem. With a freshly installed plugin it always works, and then *something* happens and all of a sudden it doesn't.

Anyway, should at least not stop working now. I think. I hope.